Add support for folders not in the root filesystem
Telling Finder to `open folder` with the string representation of
certain HFS paths results in an error.

Prior to this commit, this would cause the script to error out for
support folders that belong to a mounted filesystem which is not the
root filesystem or the home directory.

Steps to reproduce:

- Create and import a ZFS zpool.
- Create a dataset with a different mountpoint than the
  root filesystem or the home directory.
- Mount the dataset from the zpool.
- Run the script.
- Choose a folder from inside the zpool.
- The folder will not open in Finder.

This commit fixes that issue by converting the HFS path back from a
string representation to an `alias`.

It also ensures backwards compatibility with existing support
folder entries created after commit 26d9563.